cleaned up utils/package
