+++ src/components/CityScopeJS/BaseMap/layers/RoboscopeGridLayer.js
@@ -1,5 +1,34 @@
 import { GeoJsonLayer } from "deck.gl";
-import { _handleGridcellEditing, translate, color_opc } from "../utils/BaseMapUtils";
+import { _handleGridcellEditing } from "../utils/BaseMapUtils";
+
+// Color Modifier: Opaque for selected pixels, transparent for unselected 
+export const color_opc = (color, op) => {
+  return [color[0], color[1], color[2], op]
+}
+
+// Translate Helper Functions: takes a list of indices and a start/end index to shift list of indices 
+const twoDcoords = (header, index) => {
+  return [parseInt(index%header.ncols), parseInt(index/header.ncols)]
+}
+
+const flat_index = (data, x,y) => {
+  return y*data.ncols + x
+}
+
+export const translate = (data, features, start, end) => {
+  var output = []
+  const [x1,y1] = twoDcoords(data, start);
+  const [x2,y2] = twoDcoords(data, end);
+  const [x_f, y_f] = twoDcoords(data, features[0]);
+  const [x_l, y_l] = twoDcoords(data, features[features.length-1]);
+  const change_y= (y2-y1>0) ? Math.min(y2-y1, data.nrows-y_l) : Math.max(y2-y1, 0-y_f)
+  const change_x= (x2-x1 > 0) ? Math.min(x2-x1, data.ncols-x_l) : Math.max(x2-x1, 0-x_f)  
+  features.forEach((item) => {
+    var [x, y] = twoDcoords(data, item);
+    output.push(flat_index(data, x+change_x, y+change_y));
+  });
+  return output
+}
 
 export default function GridLayer({
     data,
